Created new file for CalculatorArgParser
